requirements.txt
streamlit
google-genai
python-docx
fuzzywuzzy

